backend/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
supabase>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
